+++ gonggu-market-api/src/main/java/com/gonggu/market/api/controller/auth/AuthController.java
@@ -1,11 +1,14 @@
 package com.gonggu.market.api.controller.auth;
 
 
+import com.gonggu.market.api.config.auth.PrincipalDetails;
 import com.gonggu.market.api.domain.user.User;
 import com.gonggu.market.api.dto.auth.SignupDto;
+import com.gonggu.market.api.dto.user.UserUpdateDto;
 import com.gonggu.market.api.service.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.Cookie;
@@ -25,5 +28,14 @@ public class AuthController {
     public ResponseEntity<SignupDto> signup(@RequestBody SignupDto dto) {
         SignupDto result = authService.signup(dto);
         return ResponseEntity.ok(result);
+    }
+
+    @PutMapping("/{userId}/update")
+    public ResponseEntity<?> update(@PathVariable Long userId,
+                                    @RequestBody UserUpdateDto dto,
+                                    @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        User userEntity = authService.update(userId, dto);
+        principalDetails.setUser(userEntity);
+        return ResponseEntity.noContent().build();
     }
 }
