+++ gonggu-market-api/src/main/java/com/gonggu/market/api/service/AuthService.java
@@ -7,6 +7,7 @@ import com.gonggu.market.api.domain.address.AddressRepository;
 import com.gonggu.market.api.domain.user.User;
 import com.gonggu.market.api.domain.user.UserRepository;
 import com.gonggu.market.api.dto.auth.SignupDto;
+import com.gonggu.market.api.dto.user.UserUpdateDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -58,5 +59,28 @@ public class AuthService {
                         userEntity.getDetailAddress(),
                 userEntity.getAddress().getZipcode()
         );
+    }
+
+    public User update(Long id, UserUpdateDto dto) {
+        User userEntity = userRepository.findById(id).orElseThrow(() -> {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        });
+
+        String rawPassword = dto.getPassword();
+        String encPassword = passwordEncoder.encode(rawPassword);
+        userEntity.setPassword(encPassword);
+        userEntity.setNickname(dto.getNickname() == null ? userEntity.getNickname() : dto.getNickname());
+        userEntity.setMobile(dto.getMobile() == null ? userEntity.getMobile() : dto.getMobile());
+        userEntity.setDetailAddress(dto.getAddress() == null ? userEntity.getDetailAddress() : dto.getAddress());
+        if (dto.getZipcode() != null) {
+            Address address = addressRepository.findByZipcode(dto.getZipcode());
+            if (address == null) {
+                System.out.println("address : " + address.toString());
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
+            userEntity.setAddress(address);
+        }
+        userEntity = userRepository.save(userEntity);
+        return userEntity;
     }
 }
