+++ gonggu-market-api/src/main/java/com/gonggu/market/api/service/UserService.java
@@ -34,30 +34,19 @@ public class UserService {
         this.passwordEncoder = passwordEncoder;
     }
 
-    public User update(Long id, UserUpdateDto dto) {
-        User userEntity = userRepository.findById(id).orElseThrow(() -> {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        });
-
-        String rawPassword = dto.getPassword();
-        String encPassword = passwordEncoder.encode(rawPassword);
-        userEntity.setPassword(encPassword);
-        userEntity.setNickname(dto.getNickname() == null ? userEntity.getNickname() : dto.getNickname());
-        userEntity.setMobile(dto.getMobile() == null ? userEntity.getMobile() : dto.getMobile());
-        userEntity.setDetailAddress(dto.getAddress() == null ? userEntity.getDetailAddress() : dto.getAddress());
-        if (dto.getZipcode() != null) {
-            Address address = addressRepository.findByZipcode(dto.getZipcode());
-            if (address == null) {
-                System.out.println("address : " + address.toString());
-                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-            }
-            userEntity.setAddress(address);
-        }
-        userEntity = userRepository.save(userEntity);
-        return userEntity;
+    public UserProfileDto profile(String userId) {
+        User userEntity = userRepository.findById(Long.valueOf(userId)).get();
+        return new UserProfileDto(
+                userEntity.getEmail(),
+                userEntity.getNickname(),
+                userEntity.getMobile(),
+                userEntity.getDetailAddress(),
+                userEntity.getAddress().getZipcode(),
+                userEntity.getPoint()
+        );
     }
 
-    public UserProfileDto profile(String userId, String token) {
+    public UserProfileDto detail(String userId, String token) {
         token = token.replace(JwtProperties.TOKEN_PREFIX, "");
         Long userIdFromToken = JWT.decode(token).getClaim("id").asLong();
         logger.info(userIdFromToken.toString());
